requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
firebase-admin==6.5.0
python-dateutil==2.9.0.post0
requests==2.31.0
websockets==12.0
gunicorn==22.0.0